refactor(api): move version-repository file from certification evaluation to shared
