+++ api/tests/certification/shared/integration/infrastructure/repositories/version-repository_test.js
@@ -1,6 +1,6 @@
-import { Version } from '../../../../../../src/certification/evaluation/domain/models/Version.js';
-import * as versionRepository from '../../../../../../src/certification/evaluation/infrastructure/repositories/version-repository.js';
 import { Frameworks } from '../../../../../../src/certification/shared/domain/models/Frameworks.js';
+import { Version } from '../../../../../../src/certification/shared/domain/models/Version.js';
+import * as versionRepository from '../../../../../../src/certification/shared/infrastructure/repositories/version-repository.js';
 import { NotFoundError } from '../../../../../../src/shared/domain/errors.js';
 import { catchErr, databaseBuilder, expect } from '../../../../../test-helper.js';
