+++ api/tests/certification/shared/unit/domain/models/Version_test.js
@@ -1,5 +1,5 @@
-import { Version } from '../../../../../../src/certification/evaluation/domain/models/Version.js';
 import { Frameworks } from '../../../../../../src/certification/shared/domain/models/Frameworks.js';
+import { Version } from '../../../../../../src/certification/shared/domain/models/Version.js';
 import { EntityValidationError } from '../../../../../../src/shared/domain/errors.js';
 import { catchErrSync, expect } from '../../../../../test-helper.js';
