+++ api/tests/tooling/domain-builder/factory/certification/evaluation/build-version.js
@@ -1,5 +1,5 @@
-import { Version } from '../../../../../../src/certification/evaluation/domain/models/Version.js';
 import { Frameworks } from '../../../../../../src/certification/shared/domain/models/Frameworks.js';
+import { Version } from '../../../../../../src/certification/shared/domain/models/Version.js';
 
 export const buildVersion = ({ id = 1, scope = Frameworks.CORE, challengesConfiguration = {} } = {}) => {
   return new Version({
